feat: アイコンパイプラインを ResultsView へ移設(#646 PR2・決定 5)
TextureHandle は窓の ctx 従属のため、行描画と同じ窓に置く。
